+++ app/models/maths_and_physics/slug_sequence.rb
@@ -12,6 +12,19 @@ module MathsAndPhysics
     SLUGS = [
       "teaching-maths-or-physics",
       "eligibility-confirmed",
+      "information-provided",
+      "verified",
+      "address",
+      "gender",
+      "teacher-reference-number",
+      "national-insurance-number",
+      "student-loan",
+      "student-loan-country",
+      "student-loan-how-many-courses",
+      "student-loan-start-date",
+      "email-address",
+      "bank-details",
+      "check-your-answers",
       "ineligible",
     ].freeze
 
@@ -23,6 +36,11 @@ module MathsAndPhysics
 
     def slugs
       SLUGS.dup.tap do |sequence|
+        sequence.delete("student-loan-country") if claim.no_student_loan?
+        sequence.delete("student-loan-how-many-courses") if claim.no_student_loan? || claim.student_loan_country_with_one_plan?
+        sequence.delete("student-loan-start-date") if claim.no_student_loan? || claim.student_loan_country_with_one_plan?
+        sequence.delete("address") if claim.address_verified?
+        sequence.delete("gender") if claim.payroll_gender_verified?
       end
     end
   end